# MINOR: Add `KafkaAdminClient.getListOffsetsCalls` benchmark