--- a/clients/src/test/java/org/apache/kafka/clients/admin/AdminClientTestUtils.java
+++ b/clients/src/test/java/org/apache/kafka/clients/admin/AdminClientTestUtils.java
@@ -110,9 +110,9 @@ public class AdminClientTestUtils {
      * from within the admin package.
      */
     public static List<KafkaAdminClient.Call> getListOffsetsCalls(KafkaAdminClient adminClient, 
-                                                   MetadataOperationContext<ListOffsetsResult.ListOffsetsResultInfo, ListOffsetsOptions> context,
-                                                   Map<TopicPartition, OffsetSpec> topicPartitionOffsets,
-                                                   Map<TopicPartition, KafkaFutureImpl<ListOffsetsResult.ListOffsetsResultInfo>> futures) {
+                                                                  MetadataOperationContext<ListOffsetsResult.ListOffsetsResultInfo, ListOffsetsOptions> context,
+                                                                  Map<TopicPartition, OffsetSpec> topicPartitionOffsets,
+                                                                  Map<TopicPartition, KafkaFutureImpl<ListOffsetsResult.ListOffsetsResultInfo>> futures) {
         return adminClient.getListOffsetsCalls(context, topicPartitionOffsets, futures); 
     }
 }