new benchmark to test KafkaAdminClient.getListOffsetsCalls
originated from https://github.com/apache/kafka/pull/10940

### Committer Checklist (excluded from commit message)
- [ ] Verify design and implementation 
- [ ] Verify test coverage and CI build status
- [ ] Verify documentation (including upgrade notes)