@jeffkbkim 2021-07-02
ran benchmark with warm up = 5, iteration = 5

With changes:
```
Benchmark                                             (partitionCount)  (topicCount)  Mode  Cnt       Score       Error  Units
GetListOffsetsCallsBenchmark.testGetListOffsetsCalls               100             1  avgt    5      40.040 ±     3.506  us/op
GetListOffsetsCallsBenchmark.testGetListOffsetsCalls               100            10  avgt    5     474.825 ±    95.547  us/op
GetListOffsetsCallsBenchmark.testGetListOffsetsCalls              1000             1  avgt    5     429.767 ±    60.605  us/op
GetListOffsetsCallsBenchmark.testGetListOffsetsCalls              1000            10  avgt    5    5069.417 ±   662.226  us/op
GetListOffsetsCallsBenchmark.testGetListOffsetsCalls             10000             1  avgt    5    6000.379 ±   235.654  us/op
GetListOffsetsCallsBenchmark.testGetListOffsetsCalls             10000            10  avgt    5  121240.173 ± 49894.963  us/op
```

Without changes:
```
GetListOffsetsCallsBenchmark.testGetListOffsetsCalls partition=100,       topic=1: Iteration   5: 4860.989 us/op
GetListOffsetsCallsBenchmark.testGetListOffsetsCalls partition=100,     topic=10: Iteration   1: 424064.225 us/op
GetListOffsetsCallsBenchmark.testGetListOffsetsCalls partition=1000,     topic=1: Iteration   5: 452289.400 us/op
GetListOffsetsCallsBenchmark.testGetListOffsetsCalls partition=1000,   topic=10: Iteration   2: 49306777.392 us/op
GetListOffsetsCallsBenchmark.testGetListOffsetsCalls partition=10000,   topic=1: Iteration   5: 53940393.545 us/op
GetListOffsetsCallsBenchmark.testGetListOffsetsCalls partition=10000, topic=10: Warm Up  1: 14664107350.721 us/op
```
Running the benchmark pre-https://github.com/apache/kafka/pull/10940 took too long to finish. included the median result for each topic-partition pair. case for partition=10000, topic=10 took too long during warm up (each take +4 hours).