@dajac 2021-07-05
@jeffkbkim We could add a `getListOffsetsCalls` method to `AdminClientTestUtils`, which is in the same package as the `KafkaAdminClient`, to access the package private method from the benchmark.